Tests/LibWeb/Text/expected/html-parser-foster-parented-insertion-is-counted.txt

a foster-parented insertion computed an explicit offset
